Ci - Add ChangeSet (#3)
This PR is focused on the implementation and subsequent modifications
regarding the use of @changesets/cli in our project. This work was
carried out through a series of steps, defined by the following commits
in chronological order:

- feat(pnpm-lock.yaml): Add multiple packages related to
'@changesets/cli' The @changesets/cli package was introduced into the
project, along with related packages, updating the pnpm-lock.yaml and
package.json.

- chore(changeset): changeset init Initial setup files and documentation
for @changesets/cli were created, helping developers to effectively
version and publish their code.

- ci(github-workflows): add changesets requirement check to PR workflow
A requirement for changesets was added to the pull request workflow,
which enforces that no PR can be merged into main without a
corresponding changeset.

- ci(.github): Add Release workflow A new GitHub Action workflow was
introduced which sets up Node.js with pnpm, manages dependencies
installation as well as caching and release automation.

- chore(package.json): add changeset scripts Three new scripts were
added to package.json to enable convenient use of the 'changeset'
command.

- chore(changeset): Set up changeset cli and CI The @changesets/cli
build tool was further configured for versioning and publishing, with
added GitHub Action workflows and updated pull request checks.

- fix(pull-request-checks-workflow): update changeset status reference
The changeset command was updated in the pull-request checks workflow
with reference changed from 'main' to 'origin/main'.

Through the implementation of @changesets/cli and the related
modifications, we are expecting to handle package versioning more
effectively and maintain detailed, autogenerated changelogs. This will
contribute to the preservation of precise versioning and release
information throughout the project's lifecycle.

---------

Signed-off-by: Giovanni Ravalico <15946771+suddenlyGiovanni@users.noreply.github.com>