feat: remove `resume.json`
The resume.json file that was present in the root directory of the
resume module has been entirely deleted. This file consisted of
extensive personal and professional information. It is no longer
required, hence removed from the project module.

Signed-off-by: Giovanni Ravalico <15946771+suddenlyGiovanni@users.noreply.github.com>

chore: update package manager version and lockfile

Updated the `pnpm` package manager version from `8.15.6` to `9.0.6` in
the `package.json` file and modified the lockfile structure in
`pnpm-lock.yaml`. The lockfile version was also updated to reflect the
changes. Changes in the lockfile include updates to devDependencies
format, packages list, and resolution integrity values.

Signed-off-by: Giovanni Ravalico <15946771+suddenlyGiovanni@users.noreply.github.com>

chore(🦋): add changeset

Signed-off-by: Giovanni Ravalico <15946771+suddenlyGiovanni@users.noreply.github.com>