feat(resume): refactor resume schema and data structure
This commit refactors the resume.yml and schema.json files to improve
structure and readability. It separates roles, responsibilities, and
technology stack for each work experience, separating key details to
create a clearer overview of each position held. It also updates the
schema to ensure data's validity and reflects the changes made to the
resume structure.

Signed-off-by: Giovanni Ravalico <15946771+suddenlyGiovanni@users.noreply.github.com>