feat(resume-schema): remove startDate and endDate fields work
This commit removes the startDate and endDate fields from the work
module in the resume-schema. These fields were previously used to
indicate the starting and ending dates of employment at a particular
company. Their removal simplifies the schema and focuses on the roles
and location of work. Said info can now be derived/computed from roles
start and end date

Signed-off-by: Giovanni Ravalico <15946771+suddenlyGiovanni@users.noreply.github.com>