refactor(resume): update job responsibilities and highlights
This commit refines the job responsibilities and achievements sections
in the resume file. It includes more specific details about maintaining
the customer-facing GraphQL API and CI/CD pipelines. Also, it clarifies
the role in the re-architecture of the backend system using
Domain-Driven Design principles and elucidates accomplishments related
to using RFC and ADR processes for team decision-making.

Signed-off-by: Giovanni Ravalico <15946771+suddenlyGiovanni@users.noreply.github.com>