feat(resume): update Software Developer role responsibilities and highlights
The responsibilities and highlights of the Software Developer role
were thoroughly revised for clarity. Specific contributions to the
maintenance and improvement of the Exercise App's codebase, the
refactor of the compiler, and the promotion of process improvements
were added. Related highlights were reorganized and conveyed in a
structured manner to enhance readability and usability.

Signed-off-by: Giovanni Ravalico <15946771+suddenlyGiovanni@users.noreply.github.com>