Renovate/workspace deps (#161)
This pull request includes several changes to remove the use of
`changesets` and update dependencies across the project. The most
important changes include the deletion of `changesets` configuration
files and tasks, and the update of `vitest` and other dependencies.

Removal of `changesets`:

*
[`.changeset/README.md`](diffhunk://#diff-a34534bc93c6a37355abf0bdd229cb9fa11a8b3a2de4ca2d97e71356d361b459L1-L8):
Deleted the entire file, which provided documentation for `changesets`.
*
[`.changeset/config.json`](diffhunk://#diff-ed5b41335ff968c5cfd8035d4b8f8c8c0538b6746182d522adb0312eb9137fc5L1-L10):
Deleted the configuration file for `changesets`.
*
[`deno.json`](diffhunk://#diff-54337bcff4b919f6b8b4e0b0e51d06fd9124562fde88899be726c24b7e354b7cL34-R43):
Removed all `changeset` tasks and related descriptions.
*
[`.github/workflows/release.yml`](diffhunk://#diff-87db21a973eed4fef5f32b267aa60fcee5cbdf03c67fafdc2a9b553bb0b15f34L4-L6):
Removed the `push` trigger for the release workflow, which was
previously set to the `main` branch.

Dependency updates:

* `deno.json`, `packages/resume/deno.json`,
`packages/schema-resume/deno.json`: Updated the version of `vitest` from
`3.0.4` to `3.0.5` and other dependencies to their latest versions.
[[1]](diffhunk://#diff-54337bcff4b919f6b8b4e0b0e51d06fd9124562fde88899be726c24b7e354b7cL34-R43)
[[2]](diffhunk://#diff-68285fbdc064485a9ee7acb4919b7bf6b79a928825fa2554198415ffd60901e8L9-R9)
[[3]](diffhunk://#diff-93010554af580a51e0400e18e885ef3149a109636c486d56ebf2eac14ff61567L16-R16)
[[4]](diffhunk://#diff-93010554af580a51e0400e18e885ef3149a109636c486d56ebf2eac14ff61567L40-R40)

---------

Signed-off-by: suddenlyGiovanni <15946771+suddenlyGiovanni@users.noreply.github.com>